refactor: Add module_pw to cmake fix: a bug of Chebyshev that it is wrong when using float

+++ source/module_base/math_chebyshev.h
@@ -212,7 +212,7 @@ public:
 	FFTW(const int norder2_in);
 	~FFTW();
 	void execute_fftw();
-    double* dcoef; //[norder2]
+    float* dcoef; //[norder2]
 	fftwf_complex *ccoef;
 	fftwf_plan coef_plan;
 };
